Add blog and project pages
Add a Distill blog. Remove the author metadata on the main listing page.

Add a page for projects. Apply a custom CSS to create card container,
adopting the stylesheet from W3Schools's column cards. Add Indonesia
EconRecovery and jakpoststyle as initial two projects on the page. This
page is meant to serve as a portfolio gallery.

Add several more items to the YAML, including possible base and
repo URL, more menus to the navbar, options to share blog posts,
license and Twitter handle. Also, changes the website title to make
it more personal